+++ src/index.js
@@ -1,4 +1,10 @@
-export { GeoGlobe, createGlobe, world } from "./geo-globe.js";
+export {
+  GeoGlobe,
+  GeoGlobe as CanvasGlobe,
+  createGlobe,
+  createGlobe as createCanvasGlobe,
+  world,
+} from "./geo-globe.js";
 export { themes, countryPalette } from "./themes.js";
 export { presets } from "./presets.js";
 export { scenes } from "./scenes.js";
@@ -8,5 +14,6 @@ export { locateViewer, locateViewerPrecise, timeZoneLocation, countryLocation, p
 export { recordCanvas, downloadBlob, canRecord, supportedRecordingType } from "./recorder.js";
 export { SphereTexture } from "./texture.js";
 export { Media, drawFitted } from "./media.js";
+export { DEFAULT_LICENSE_KEY, OPEN_SOURCE_LICENSE_KEY, inspectLicenseKey, hasLicenseKey } from "./license.js";
 export { mapAspect, colorScale, subsolarPoint, greatCircle, angularDistance, pointInGeometry, geometryBounds, projections } from "./geo.js";
 export { default } from "./geo-globe.js";